Address shellcheck warnings (#510)
* non-functional, remove trailing spaces (#498)


* Fix SC2155 of shellchecker (#498)

SC2155: Declare and assign separately to avoid masking return values.
See https://github.com/koalaman/shellcheck/wiki/SC2155

Changed all instances of "local" and "export"


* Fix SC2027 of shellchecker (#498)

SC2027: The surrounding quotes actually unquote this. Remove or escape them.
See https://github.com/koalaman/shellcheck/wiki/SC2027

Changed single occurence


* Make build.sh executable (#498)


* Fix SC2179 of shellcheck (#498)

Use array+=("item") to append items to an array.
See https://github.com/koalaman/shellcheck/wiki/SC2179


* Fix SC2006 of shellchecker (#498)

Use $(...) notation instead of legacy backticked `...`.
See https://github.com/koalaman/shellcheck/wiki/SC2006


* Fix SC2219 of shellchecker (#498)

Instead of let expr, prefer (( expr )) .
See https://github.com/koalaman/shellcheck/wiki/SC2219


* Fix SC2166 of shellchecker (#498)

SC2166: Prefer [ p ] || [ q ] as [ p -o q ] is not well defined.
See https://github.com/koalaman/shellcheck/wiki/SC2166


* Remove duplicate code (#498)


* Fix SC2154 of shellchecker (#498)

SC2154: var is referenced but not assigned.
See https://github.com/koalaman/shellcheck/wiki/SC2154

This part of the code does not seem finished.
Remove the sed's for the variables which don't exist, but keep the
original code as comment. When the relevant code is fixed the
commented code may be used again.


* Fix SC2091 of shellchecker (#498)

SC2091: Remove surrounding $() to avoid executing output.
See https://github.com/koalaman/shellcheck/wiki/SC2091


* Fix some SC2086 of shellchecker (#498)

SC2086: Double quote to prevent globbing and word splitting.
See https://github.com/koalaman/shellcheck/wiki/SC2086

This only fixes very few instances of this occurence.
There are 147 more occurences which are not fixed by this commit.


* Fix SC2115 of shellchecker (#498)

SC2115: Use "${var:?}" to ensure this never expands to /* .
See https://github.com/koalaman/shellcheck/wiki/SC2115


* non-functional, improve indention (#498)

The indention was changed so it is everywhere 2 spaces.
This is according to the decission done at
https://github.com/openhab/openhabian/issues/498#issuecomment-451700319

+++ functions/java-jre.sh
@@ -18,27 +18,30 @@ java_webupd8_archive() {
 }
 
 java_zulu() {
-  local FILE="/var/tmp/.zulu.$$"
-  local INSTALLROOT=/opt/jdk
-  local TEMPROOT=/opt/jdk-new
+  local FILE
+  local INSTALLROOT
+  local TEMPROOT
+  local JAVA
+  FILE="/var/tmp/.zulu.$$"
+  INSTALLROOT=/opt/jdk
+  TEMPROOT=/opt/jdk-new
   mkdir ${INSTALLROOT}
   mkdir ${TEMPROOT}
-  
+
   if is_arm; then
     # Latest version check https://www.azul.com/downloads/zulu-embedded
-    local JAVA=zulu8.33.0.134-jdk1.8.0_192-linux_aarch32hf
+    JAVA=zulu8.33.0.134-jdk1.8.0_192-linux_aarch32hf
   else
     # Latest version check https://www.azul.com/downloads/zulu-linux
-    local JAVA=zulu8.33.0.1-jdk8.0.192-linux_x64
+    JAVA=zulu8.33.0.1-jdk8.0.192-linux_x64
   fi
   whiptail --textbox $BASEDIR/includes/azul_zulu_license.md --scrolltext 27 116
   
   cond_redirect wget -nv -O $FILE http://cdn.azul.com/zulu-embedded/bin/${JAVA}.tar.gz
   cond_redirect tar -xpzf $FILE -C ${TEMPROOT}
   if [ $? -ne 0 ]; then echo "FAILED (Zulu java)"; rm -f ${FILE}; exit 1; fi
-  rm -rf $FILE ${INSTALLROOT}/*
+  rm -rf $FILE ${INSTALLROOT:?}/*
   mv ${TEMPROOT}/* ${INSTALLROOT}/; rmdir ${TEMPROOT}
   cond_redirect update-alternatives --install /usr/bin/java java ${INSTALLROOT}/${JAVA}/bin/java 1083000
   cond_redirect update-alternatives --install /usr/bin/javac java ${INSTALLROOT}/${JAVA}/bin/javac 1083000
 }
-
